Refactor application into five microservices with Nginx routing
- Decoupled the original Go application into five services:
  - api_get_books
  - api_post_books
  - api_put_books
  - api_delete_books
  - frontend_renderer
- Each service has its own multi-stage Dockerfile for amd64.
- Configured Nginx to route traffic based on HTTP method for /api/books
  and path for other requests (frontend, static assets).
- Updated docker-compose.yml to orchestrate all services, including MongoDB.
- Ensured MONGO_URI is passed as an environment variable.
- Copied views/ and css/ to frontend_renderer service for its build context.
- Removed original monolithic cmd/main.go and root Dockerfile.

Note: Runtime testing was blocked by Docker daemon permission issues
in the execution environment.
