+++ services/frontend_renderer/.gitkeep
@@ -1,0 +1,1 @@
+# This file is a placeholder to ensure the directory is created.
